+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/theme/ElementProvider.java
@@ -44,7 +44,7 @@ public class ElementProvider implements IElementProvider
         if (element instanceof SecuritiesChart securitiesChart)
             return new SecuritiesChartElementAdapter(securitiesChart, engine);
         if (element instanceof PortfolioBalanceChart portfolioBalanceChart)
-            return new PortfolioBalanceChartElementAdapter(portfolioBalanceChart, engine);
+            return new PortfolioBalanceChartCSS.ElementAdapter(portfolioBalanceChart, engine);
         if (element instanceof ValueColorScheme scheme)
             return new ValueColorSchemeElementAdapter(scheme, engine);
